+++ .github/workflows/test-all-versions.pr.yml
@@ -16,7 +16,7 @@ jobs:
     env:
       PR_LABELS: ${{ toJSON(github.event.pull_request.labels.*.name) }}
     outputs:
-      args: ${{ steps.lerna-args.outputs.args }}
+      args: ${{ steps.npm-workspace-args.outputs.args }}
     steps:
       - name: Checkout
         uses: actions/checkout@v4
@@ -24,14 +24,14 @@ jobs:
       - name: Install lerna
         run: npm install -g lerna@6.6.2
       - name: Parse labels into lerna scope arguments
-        id: lerna-args
+        id: npm-workspace-args
         run: |
-          OUTPUT=`node scripts/parse-lerna-scopes.mjs "$PR_LABELS"`
+          OUTPUT=`node scripts/pr-labels-to-npm-workspace-args.mjs "$PR_LABELS"`
           echo "args=$OUTPUT" >> $GITHUB_OUTPUT
 
   tav:
     uses: ./.github/workflows/test-all-versions.yml
     needs: parse-labels
     with:
-      lerna-args: ${{ needs.parse-labels.outputs.args }}
+      npm-workspace-args: ${{ needs.parse-labels.outputs.args }}
     if: ${{ needs.parse-labels.outputs.args != '' }}
