+++ plugins/node/opentelemetry-instrumentation-aws-sdk/.tav.yml
@@ -1,16 +1,18 @@
 "aws-sdk":
-  # there are so many version to test, it can take forever.
-  # we will just sample few of them
-  versions: ">=2.1266.0 || 2.1262.0 || 2.1219.0 || 2.1048.0 || 2.1012.0 || 2.647.0 || 2.308.0"
+  # A small subset of releases in the range [2.308.0, 3) to reduce testing time.
+  versions: "2.308.0 || 2.548.0 || 2.785.0 || 2.1025.0 || 2.1265.0 || 2.1506.0 || >=2.1508.0"
   commands:
     - npm run test
 
 "@aws-sdk/client-s3":
-  versions: ">=3.223.0 || 3.218.0 || 3.216.0 || 3.154.0 || 3.107.0 || 3.54.0 || 3.6.1"
+  # A small subset of releases in the range [3.6.1, 4) to reduce testing time.
+  # - 3.377.0 was a bad release (see issue #1828).
+  versions: "3.6.1 || 3.53.0 || 3.163.0 || 3.266.0 || 3.354.0 || 3.458.0 || >=3.462.0"
   commands:
     - npm run test
 
 "@aws-sdk/client-sqs":
-  versions: ">=3.216.0 || 3.171.0 || 3.58.0 || 3.54.0 || 3.43.0 || 3.24.0"
+  # A small subset of releases in the range [3.24.0, 4) to reduce testing time.
+  versions: "3.24.0 || 3.85.0 || 3.194.0 || 3.278.0 || 3.357.0 || 3.461.0 || >=3.462.0"
   commands:
     - npm run test
